Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,6 +5,6 @@ int n1,n2,n3,n4;
 float n5;
 scanf("%d%d%d%d%d",&n1,&n2,&n3,&n4,&n5);
 n4=(n1+n2+n3);
-n5=n4/3;
+n5=(float)n4/3;
 printf("Average: %.2f",n5);
 }
